refactor shark
