+++ src/index.ejs
@@ -13,7 +13,14 @@
         }
 
         #renderCanvas {
-            width   : 100%;
+            width   : 50%;
+            height  : 100%;
+            touch-action: none;
+        }
+        #renderCanvas2 {
+            position: absolute;
+            left: 50%;
+            width   : 50%;
             height  : 100%;
             touch-action: none;
         }
@@ -21,5 +28,6 @@
 </head>
 <body>
     <canvas id="renderCanvas"></canvas>
+    <canvas id="renderCanvas2"></canvas>
 </body>
 </html>
